+++ src/cliprelay/qml/Theme.qml
@@ -91,7 +91,7 @@ QtObject {
     readonly property int workspaceTabHeight: 34
     readonly property int prepareStudioHeaderHeight: 38
     readonly property int prepareInspectorTabsHeight: 40
-    readonly property int prepareSourceHeight: 62
+    readonly property int prepareSourceHeight: 48
     readonly property int radiusWorkbench: 4
     readonly property int focusWidth: 2